feat: define Product and Service interfaces, update related components and hooks
