+++ src/components/ActiveReservationPage/ProductsCard.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Product } from "@/types/reservation";
+import { Product } from "@/types/product";
 
 interface ProductsCardProps {
   products: Product[];
